+++ tests/testthat/test-category_table.R
@@ -2,7 +2,7 @@ test_that("x", {
   marks <- data.frame(ID = 7, K = 1, E = 2, C = 3)
   cols <- c(K = "Knowledge", E = "Evaluation", C = "Communication")
   cats <- c("1" = "Bad", "2" = "OK", "3" = "Good")
-  symbol <- "*"
+  symbol <- "&#10004;"
   x <- category_table(marks, cols, cats, symbol, F)
   x2 <- data.frame(
     Criteria = factor(cols),
@@ -18,7 +18,7 @@ test_that("case insensitivity", {
   marks <- data.frame(ID = 7, K = "Good", E = "good", C = "Bad")
   cols <- c(K = "Knowledge", E = "Evaluation", C = "Communication")
   cats <- c("Bad", "OK", "Good")
-  symbol <- "*"
+  symbol <- "&#10004;"
   x <- category_table(marks, cols, cats, symbol, F)
   x2 <- data.frame(
     Criteria = factor(cols),
@@ -34,7 +34,7 @@ test_that("no cats", {
   # numeric cats
   marks <- data.frame(ID = 7, K = 1, E = 2, C = 3)
   cols <- c("K", "C", "E")
-  symbol <- "*"
+  symbol <- "&#10004;"
   x <- category_table(marks, cols, kable = F) %>% tibble::as_tibble()
   # data frame is weird with integer column headers
   x2 <- tibble::tibble(
@@ -49,7 +49,7 @@ test_that("no cats", {
   # character cats
   marks <- data.frame(ID = 7, K = "A", E = "B", C = "C")
   cols <- c("K", "C", "E")
-  symbol <- "*"
+  symbol <- "&#10004;"
   x <- category_table(marks, cols, kable = F)
   x2 <- data.frame(
     Criteria = factor(cols),